- unit-based time

+++ Game/src/com/philon/rpg/util/TimerListObject.java
@@ -3,16 +3,12 @@ package com.philon.rpg.util;
 import com.philon.rpg.map.mo.RpgMapObj;
 
 public class TimerListObject {
-	public int timerValue;
+	public float timerValue;
 	public RpgMapObj mo;
 
-	//----------
-	
-	public TimerListObject( RpgMapObj newMo, int newTimerValue ) {
+	public TimerListObject( RpgMapObj newMo, float newTimerValue ) {
 		mo = newMo;
 		timerValue = newTimerValue;
 	}
-
-	//----------
 
 }
